draw_filters moved back to PostProcessingController but enum class moved as well to eliminate app reliance
