+++ engine/include/engine/graphics/PostProcessingController.hpp
@@ -3,6 +3,15 @@
 
 #include <engine/core/Controller.hpp>
 
+enum class Filter : int {
+  NONE = 0,
+  GRAYSCALE = 1,
+  NEGATIVE = 2,
+  BLACKWHITE = 3,
+  DEEPFRIED = 4,
+  OUTLINE = 5
+};
+
 namespace engine::graphics {
   class PostProcessingController : public engine::core::Controller {
     unsigned int m_hdrFBO = 0;
@@ -11,6 +20,8 @@ namespace engine::graphics {
     unsigned int m_pingpongBuffer[2] = { 0, 0 };
     unsigned int m_screenFBO = 0;
     unsigned int m_screenTexture = 0;
+
+    Filter m_active_filter = Filter::NONE;
 
   public:
     std::string_view name() const override {
@@ -22,7 +33,8 @@ namespace engine::graphics {
     void prepare_bloom_effect(unsigned int wHeight, unsigned int wWidth);
     void prepare_bloom_shaders();
 
-    unsigned int get_screen_texture();
+    Filter get_active_filter() const;
+    void set_active_filter(Filter filter);
 
     void draw() override;
 
@@ -30,9 +42,9 @@ namespace engine::graphics {
     void end_draw() override;
 
     void draw_bloom();
+    void draw_filter();
 
     void terminate() override;
-
   };
 }
