+++ engine/src/PostProcessingController.cpp
@@ -20,8 +20,8 @@ namespace engine::graphics {
     }
 
     void PostProcessingController::draw() {
-        // I know it's a passthrough but that's on purpose so that if any other effects are added it's easier to implement
         draw_bloom();
+        draw_filter();
     }
 
     void PostProcessingController::prepare_bloom_effect(unsigned int wHeight, unsigned int wWidth) {
@@ -112,10 +112,23 @@ namespace engine::graphics {
         engine::graphics::Framebuffer::activate_texture(m_colorBuffers[0], 0);
         engine::graphics::Framebuffer::activate_texture(m_pingpongBuffer[!horizontal], 1);
         bloom_final->set_int("bloom", true);
-        bloom_final->set_float("exposure", 0.2f);
+        bloom_final->set_float("exposure", 1.0f);
         engine::graphics::Framebuffer::render_quad();
 
         engine::graphics::Framebuffer::bind_framebuffer(0);
+    }
+
+    void PostProcessingController::draw_filter() {
+        auto resources = engine::core::Controller::get<engine::resources::ResourcesController>();
+        engine::resources::Shader* post_processing_shader = resources->shader("postProcessing");
+        post_processing_shader->use();
+        post_processing_shader->set_int("effectType", static_cast<int>(m_active_filter));
+
+        engine::graphics::Framebuffer::bind_framebuffer(0);
+        engine::graphics::OpenGL::clear_buffers(false);
+        engine::graphics::Framebuffer::activate_texture(m_screenTexture, 0);
+
+        engine::graphics::Framebuffer::render_quad();
     }
 
     void PostProcessingController::terminate() {
@@ -149,7 +162,11 @@ namespace engine::graphics {
         engine::graphics::Framebuffer::delete_quad();
     }
 
-    unsigned int PostProcessingController::get_screen_texture() {
-        return m_screenTexture;
+    Filter PostProcessingController::get_active_filter() const {
+        return m_active_filter;
+    }
+
+    void PostProcessingController::set_active_filter(Filter filter) {
+        m_active_filter = filter;
     }
 }
